fix: adjust folder structure - do not rely on next.js api router. use middleware instead
